Move compile_inner into tracing/compiler.py replay_trace
